Almost complete extracting remaining registerations to modules.
