+++ src/main/java/org/spongepowered/common/data/property/store/block/IndirectlyPoweredPropertyStore.java
@@ -33,6 +33,7 @@ import org.spongepowered.api.world.World;
 import org.spongepowered.api.world.extent.Extent;
 import org.spongepowered.common.data.property.store.common.AbstractSpongePropertyStore;
 import org.spongepowered.common.registry.SpongeGameRegistry;
+import org.spongepowered.common.registry.provider.DirectionFacingProvider;
 import org.spongepowered.common.util.VecHelper;
 
 import java.util.Optional;
@@ -63,7 +64,7 @@ public class IndirectlyPoweredPropertyStore extends AbstractSpongePropertyStore<
     @Override
     public Optional<IndirectlyPoweredProperty> getFor(Location<World> location, Direction direction) {
         final net.minecraft.world.World world = (net.minecraft.world.World) location.getExtent();
-        final EnumFacing facing = SpongeGameRegistry.directionMap.get(direction);
+        final EnumFacing facing = DirectionFacingProvider.getInstance().get(direction).get();
         final boolean powered = world.getRedstonePower(VecHelper.toBlockPos(location.getBlockPosition()), facing) > 0;
         return Optional.of(new IndirectlyPoweredProperty(powered));
     }
