+++ src/main/java/org/spongepowered/common/registry/CustomCatalogRegistration.java
@@ -1,0 +1,12 @@
+package org.spongepowered.common.registry;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface CustomCatalogRegistration {
+
+}
