+++ src/main/java/org/spongepowered/common/registry/type/CookedFishRegistryModule.java
@@ -4,8 +4,9 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.collect.ImmutableList;
 import net.minecraft.item.ItemFishFood;
-import org.spongepowered.api.data.type.Fish;
-import org.spongepowered.api.data.type.Fishes;
+import org.spongepowered.api.data.type.CookedFish;
+import org.spongepowered.api.data.type.CookedFishes;
+import org.spongepowered.common.data.type.SpongeCookedFish;
 import org.spongepowered.common.registry.AdditionalRegistration;
 import org.spongepowered.common.registry.CatalogRegistryModule;
 import org.spongepowered.common.registry.RegisterCatalog;
@@ -15,33 +16,37 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
-public class FishRegistryModule implements CatalogRegistryModule<Fish> {
+public class CookedFishRegistryModule implements CatalogRegistryModule<CookedFish> {
 
-    @RegisterCatalog(Fishes.class)
-    private final Map<String, Fish> fishMap = new HashMap<>();
+    @RegisterCatalog(CookedFishes.class)
+    private final Map<String, CookedFish> fishMap = new HashMap<>();
 
     @Override
-    public Optional<Fish> getById(String id) {
+    public Optional<CookedFish> getById(String id) {
         return Optional.ofNullable(this.fishMap.get(checkNotNull(id).toLowerCase()));
     }
 
     @Override
-    public Collection<Fish> getAll() {
+    public Collection<CookedFish> getAll() {
        return ImmutableList.copyOf(this.fishMap.values());
     }
 
     @Override
     public void registerDefaults() {
         for (ItemFishFood.FishType fishType : ItemFishFood.FishType.values()) {
-            this.fishMap.put(fishType.name().toLowerCase(), (Fish) (Object) fishType);
+            if (fishType.canCook()) {
+                CookedFish cooked = new SpongeCookedFish(fishType.name(), fishType.name(), fishType);
+                this.fishMap.put(cooked.getId().toLowerCase(), cooked);
+            }
         }
     }
 
     @AdditionalRegistration
     public void registerAdditional() {
         for (ItemFishFood.FishType fishType : ItemFishFood.FishType.values()) {
-            if (!this.fishMap.containsValue((Fish) (Object) fishType)) {
-                this.fishMap.put(fishType.name().toLowerCase(), (Fish) (Object) fishType);
+            if (fishType.canCook() && !this.fishMap.containsKey(fishType.name().toLowerCase())) {
+                CookedFish cooked = new SpongeCookedFish(fishType.name(), fishType.name(), fishType);
+                this.fishMap.put(cooked.getId().toLowerCase(), cooked);
             }
         }
     }
