+++ src/main/resources/common_at.cfg
@@ -180,3 +180,5 @@ public net.minecraft.entity.monster.EntitySlime func_70799_a(I)V # setSlimeSize
 public net.minecraft.item.Item func_77656_e(I)Lnet/minecraft/item/Item; # setMaxDamage
 
 public net.minecraft.enchantment.Enchantment field_180311_a # enchantmentsList
+
+public net.minecraft.potion.Potion field_180150_I # field_180150_I
